Basic refactor + api with own http service treating promises with catching errors
